+++ client/components/app/__test__/app.action.spec.js
@@ -1,5 +1,3 @@
-import { call, put } from 'redux-saga/effects';
-import { getJson } from '../../../utils/fetch';
 import {
 	JOKE_FETCH_REQUESTED,
 	JOKE_FETCH_SUCCEEDED,
@@ -7,10 +5,9 @@ import {
 	jokeFetchRequested,
 	jokeFetchSucceeded,
 	jokeFetchFailed,
-	fetchJoke
 } from '../app.action';
 
-describe('app action creators', () => {
+describe('app.action', () => {
 	it('jokefetchRequested', () => {
 		const expectedAction = {
 			type: JOKE_FETCH_REQUESTED
@@ -41,32 +38,5 @@ describe('app action creators', () => {
 		const result = jokeFetchFailed({ message: 'network error!' });
 
 		expect(result).toEqual(expectedAction);
-	});
-});
-
-describe('sagas effects: fecthJoke', () => {
-	const iterator = fetchJoke();
-
-	it('call api', () => {
-		expect(iterator.next().value)
-			.toEqual(call(getJson, 'https://api.chucknorris.io/jokes/random'));
-	});
-
-	it('fetch success', () => {
-		expect(iterator.next().value)
-			.toEqual(put({ type: JOKE_FETCH_SUCCEEDED }));
-	});
-
-	it('fetch failed', () => {
-		const error = { message: 'unexpected error!' };
-
-		expect(iterator.throw(error).value)
-			.toEqual(put({ type: JOKE_FETCH_FAILED, message: error.message }));
-	});
-
-	it('end sagas', () => {
-		const end = { value: undefined, done: true };
-
-		expect(iterator.next()).toEqual(end);
 	});
 });
