+++ client/components/app/app.action.js
@@ -1,6 +1,3 @@
-import { call, put, takeLatest } from 'redux-saga/effects';
-import { getJson } from '../../utils/fetch';
-
 export const JOKE_FETCH_REQUESTED = 'JOKE_FETCH_REQUESTED';
 
 export const jokeFetchRequested = () => ({
@@ -20,16 +17,3 @@ export const jokeFetchFailed = error => ({
 	type: JOKE_FETCH_FAILED,
 	message: error.message || error
 });
-
-export function *fetchJoke() {
-	try {
-		const joke = yield call(getJson, 'https://api.chucknorris.io/jokes/random');
-		yield put(jokeFetchSucceeded(joke));
-	} catch (err) {
-		yield put(jokeFetchFailed(err));
-	}
-}
-
-export default function *watchFetchJoke() {
-	yield takeLatest(JOKE_FETCH_REQUESTED, fetchJoke);
-}
