+++ client/components/app/index.js
@@ -1,6 +1,3 @@
 import App from './App';
 
-export * from './app.reducer';
-export * from './app.action';
-
 export default App;
